Add rateInfo argument & fix rate display issue on touch devices
###Add fade speed as a default argument
When using a touch device rather than a mouse your finger tends to cover the text for part of the duration so its not visible long enough to read, the argument allows you to slow it down or keep it at default speed. The original behavior of the plugin is maintained if not manually overridden.

###Fix showRateInfo not showing up more than once on mobile without losing focus first
The rating display i.e. 6 / 20 is created in the mouseenter event then removed once faded out in the above duration. 
On touch devices the mouse enter event gets triggered on the initial touch which creates the text object then if the rating is clicked this is faded out and removed from the html

On subsequent touches the control doesn't lose focus and the mouse enter event doesn't get called again which means the text object is never recreated and cannot be displayed again.

This is fixed by not removing the html since this is already done in the mouseout event anyway it wasn't necessary and checking if the text object is already present but currently faded out, if so it is shown with show and then faded out immediately afterwards.

This seems to represent the same behavior as before time wise